+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add analysis configuration panel. [#31](https://github.com/IN-CORE/incore-studio/issues/31)
 - Error Handling for missing dependency graph or Valid tools in workflow. [#50](https://github.com/IN-CORE/incore-studio/issues/50)
 - Create visualization components [#18](https://github.com/IN-CORE/incore-studio/issues/18)
+- Create empty project [#42](https://github.com/IN-CORE/incore-studio/issues/42)
 
 ### Changed